:truck: Move test/koa/* to test/*
