+++ package.json
@@ -46,7 +46,7 @@
   },
   "scripts": {
     "test": "node_modules/mocha/bin/mocha -R spec --timeout=10000",
-    "test-koa": "FRAMEWORK=koa node_modules/mocha/bin/mocha -R spec --timeout=10000 test test/koa",
+    "test-koa": "FRAMEWORK=koa node_modules/mocha/bin/mocha -R spec --timeout=10000",
     "test-cov": "node_modules/mocha/bin/mocha --require blanket -R html-cov > coverage.html"
   }
 }
