+++ test/cookie-session-test.js
@@ -6,7 +6,7 @@ var request = require('supertest');
 var should = require('should');
 
 const AV = require('..');
-const appInfo = require('./utils/app-info');
+const appInfo = require('./helpers/app-info');
 
 var app = express();
