+++ test/helpers/functions.js
