+++ test/https-redirect-koa-test.js
@@ -4,8 +4,8 @@ var koa = require('koa');
 var request = require('supertest');
 require('should');
 
-var AV = require('../..');
-const appInfo = require('../utils/app-info');
+var AV = require('..');
+const appInfo = require('./helpers/app-info');
 
 var app = koa();
 
@@ -18,7 +18,7 @@ app.use(function *(next) {
 
 var server = app.listen();
 
-describe('https-redirect', function() {
+describe('https-redirect-koa', function() {
   it('should redirect', function(done) {
     request(server)
       .get('/test')
